Refactored into dedicated location picker w/o API calls
