+++ src/choose-location/choose-location.resource.ts
@@ -1,5 +1,6 @@
 import { openmrsFetch, openmrsObservableFetch } from "@openmrs/esm-api";
 import { map } from "rxjs/operators";
+import { LocationResponse } from "../types";
 
 export function getLoginLocations(): Observable<Object[]> {
   return openmrsObservableFetch(
@@ -21,41 +22,21 @@ export function setSessionLocation(
   });
 }
 
-export interface LocationResponse {
-  type: string;
-  total: number;
-  resourceType: string;
-  meta: {
-    lastUpdated: string;
-  };
-  link: Array<{
-    relation: string;
-    url: string;
-  }>;
-  id: string;
-  entry: Array<LocationEntry>;
-}
-
-export interface LocationEntry {
-  resource: Resource;
-}
-
-export interface Resource {
-  id: string;
-  name: string;
-  resourceType: string;
-  status: "active" | "inactive";
-  meta?: {
-    tag?: Array<{
-      code: string;
-      display: string;
-      system: string;
-    }>;
-  };
-}
-
-export function searchLocationsFhir(location: string) {
+export function searchLocationsFhir(
+  location: string,
+  abortController: AbortController
+) {
   return openmrsFetch<LocationResponse>(`/ws/fhir2/Location?name=${location}`, {
     method: "GET",
+    signal: abortController.signal,
   });
+}
+
+export function queryLocations(
+  location: string,
+  abortController = new AbortController()
+) {
+  return searchLocationsFhir(location, abortController).then(
+    (locs) => locs.data.entry
+  );
 }
